pyproject.toml
[project]
name = "ai-cli-runner"
version = "0.1.0"
description = "Shared async Python package for calling AI CLI tools (Claude, Gemini, Cursor) via subprocess"
requires-python = ">=3.12"
dependencies = [
    "python-simple-logger",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-xdist"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_cli_runner"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
warn_unused_configs = true
warn_redundant_casts = true
